simplify/improve app info layout & minor refactor
