+++ src/components/dialogs/app-info/sections/KeyboardShortcuts.tsx
@@ -1,5 +1,5 @@
 import { styled, Typography } from '@mui/material';
-import DialogSectionHeader from '../../../DialogSectionHeader';
+import DialogSectionHeader from '../../DialogSectionHeader';
 
 const Styled = styled('div')(() => `& {}`);
